+++ lib/drab/client.ex
@@ -43,9 +43,11 @@ defmodule Drab.Client do
                                __action: Phoenix.Controller.action_name(conn), 
                                __assigns: assigns])
       commander = controller.__drab__()[:commander]
-      modules = [Drab.Core | commander.__drab__().modules] # Drab.Core is included by default
       broadcast_topic = topic(commander.__drab__().broadcasting, controller, conn.request_path)
-      templates = Enum.map(modules, fn x -> "#{Module.split(x) |> Enum.join(".") |> String.downcase()}.js" end)
+
+      # modules = [Drab.Core | commander.__drab__().modules] # Drab.Core is included by default
+      # templates = Enum.map(modules, fn x -> "#{Module.split(x) |> Enum.join(".") |> String.downcase()}.js" end)
+      templates = Drab.all_templates_for(commander.__drab__().modules)
 
       access_session = commander.__drab__().access_session
       session = access_session